[CBRD-25617] The phenomenon that position numbers within the analytic function become incorrect occurs when the analytic function is included in the view (#5552)
http://jira.cubrid.org/browse/CBRD-25617

When a query includes analytic functions, if the columns specified in the OVER clause of the analytic function are included in the select-list, the column names are replaced with their position numbers. 

Since position numbers are used instead of column names, changing the select-list can result in different columns being associated with those position numbers.

In views, the referenced columns are prioritized when reconstructing the select-list, which can lead to issues.

A function has been added to readjust the column position numbers within the analytic function.